keep the original as current one doesnt support all models
